% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_ci.R
\name{ci_srs_wilson}
\alias{ci_srs_wilson}
\title{Estimate Wilson confidence intervals for simple random sample}
\usage{
ci_srs_wilson(
  tr0_pred0,
  tr1_pred0,
  tr0_pred1,
  tr1_pred1,
  N_sample,
  alpha = 0.05
)
}
\arguments{
\item{tr0_pred0}{A numeric value capturing the share of true negatives.}

\item{tr1_pred0}{A numeric value capturing the share of false negatives.}

\item{tr0_pred1}{A numeric value capturing the share of false positives.}

\item{tr1_pred1}{A numeric value capturing the share of true positives.}

\item{N_sample}{A numeric value capturing the total number size of the test set.}

\item{alpha}{A numeric value used capturing the type-I error rate for constructing
confidence intervals (applies to both bootstrapping and analytic CIs).}
}
\value{
A data.frame containing the upper and lower endpoints of the Wilson CIs for
the F1 score, precision and recall. These CIs are based on assuming the test set
was drawn by simple random sampling: the CIs should not be trusted if sampling
was conducted in a different way.
}
\description{
This function estimates CIs with Wilson method from confusion matrix drawn by simple random sampling
}
\details{
This function takes a confusion matrix and sample size as input, and produces
CIs with Wilson's score method for the F1 score, precision and recall. These CIs
are based on assuming the test set was drawn by simple random sampling: the CIs
should not be trusted if sampling was conducted in a different way.
}
\examples{
## enter the estimates for the test set confusion matrix cells and sample size
## get the Wilson CIs analytically under assumption that test set is a SRS
ci_srs_wilson(tr0_pred0 = 0.5, tr1_pred0 = 0.1, tr0_pred1 = 0.1, tr1_pred1 = 0.3,
N_sample = 1000, alpha = 0.05)
}
