% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_allocation.R
\name{constant_allocation}
\alias{constant_allocation}
\title{Estimate constant allocation}
\usage{
constant_allocation(data, N_sample, strata, min_per_bin = 1)
}
\arguments{
\item{data}{A data.frame from which to sample observations. One column should
contain the strata used for sampling.}

\item{N_sample}{A numeric value capturing the desired sample size.}

\item{strata}{A character value capturing the column name of the sampling strata.
Column should contain factor or character values.}

\item{min_per_bin}{A numeric value capturing the minimum number of observations per bin.
Default is 1.}
}
\value{
A named vector whose values indicate the number of observations to be
sampled per each stratum, and whose names correspond to the values in the
`strata` variable of the input dataset.
}
\description{
This function shows how many observations to sample per stratum under constant allocation
}
\details{
This function can be used to determine how many observations to sample per stratum
under constant allocation. It first divides the desired sample size equally across
all strata, and then rounds the allocated observations per bin, and lastly adjusts
them so the overall sample size is as close to the target as possible.
}
\examples{
## say that we want to sample a constant number of observations per stratum
## but with appropriate rounding given one cannot sample fractional observations.
## this function tells us how many obs per stratum to sample, given the strata
data(pop_df)
pop_df$strata <- cut(pop_df$score, 10)
constant_allocation(data=pop_df, N_sample=375, strata="strata", min_per_bin=1)
## note the function disregards NAs in the strata variable
## note the strata can be based on any information: either predicted probabilities
## (quantile- or fixed-intervals) or any other categorical information
}
