% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_metrics.R
\name{do_metrics}
\alias{do_metrics}
\title{Estimate performance metrics from confusion matrix}
\usage{
do_metrics(tr0_pred0, tr1_pred0, tr0_pred1, tr1_pred1)
}
\arguments{
\item{tr0_pred0}{A numeric value capturing the share of true negatives.}

\item{tr1_pred0}{A numeric value capturing the share of false negatives.}

\item{tr0_pred1}{A numeric value capturing the share of false positives.}

\item{tr1_pred1}{A numeric value capturing the share of true positives.}
}
\value{
A data.frame of performance metrics. The following
metrics are supported: precision, recall, F1 of positive class, accuracy, precision
of negative class, recall of negative class, F1 of negative class, MCC (Matthews
correlation coefficient), Cohen's kappa, (unweighted) macro-averaged F1 score,
weighted F1 score, BM (bookmaker's informedness). The output also shows the
proportions of true negatives, false negatives, false positives and true positives.
}
\description{
This function estimates performance metrics given a test set confusion matrix
}
\details{
This function can be used to estimate metrics from a confusion matrix. The function
can be applied regardless of how the test set was sampled, provided the estimated
share of observation in each cell of the confusion matrix is unbiased:
raw shares are unbiased if test set was SRS, stratified estimators are needed if
test set was drawn with stratified sampling. The function is called internally
within \code{\link{stratified_metrics}}. The following metrics are supported:
precision, recall, F1 score for the positive class, precision of negative class,
recall of negative class, F1 score for the negative class, accuracy, MCC (Matthews
correlation coefficient), Cohen's kappa, (unweighted) macro-averaged F1 score,
weighted F1 score, BM (bookmaker's informedness). The output also shows the
proportions of true negatives, false negatives, false positives and true positives.
}
\examples{
## enter the estimates for the test set confusion matrix cells
## get estimates of key performance metriccs
do_metrics(tr0_pred0 = 0.5, tr1_pred0 = 0.1, tr0_pred1 = 0.1, tr1_pred1 = 0.3)
}
\seealso{
\code{\link{stratified_metrics}}
}
