% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_allocation.R
\name{optimal_n_positives}
\alias{optimal_n_positives}
\title{Calculator of optimal number of positives}
\usage{
optimal_n_positives(
  N_sample = NULL,
  pi1 = NULL,
  pi0 = NULL,
  recall = NULL,
  k = NULL,
  positive_share = NULL,
  external_k = NULL,
  external_positive_share = NULL,
  weight_se_f1 = 1,
  weight_se_rec = 0,
  weight_se_prec = 0
)
}
\arguments{
\item{N_sample}{A numeric value capturing the total test set sample size considered
as binding constraint (includes both positives and negatives).}

\item{pi1}{A numeric value capturing the expected precision (share of cases
exhibiting the outcome out of predicted positives). This value must be entered
for the function to produce an output.}

\item{pi0}{A numeric value capturing the inverse of the expected precision for
the negative class (share of cases exhibiting the outcome out of predicted
negatives). If left blank, the function computes `pi0` from recall and imbalance
if `pi0` is not explicitly entered.}

\item{recall}{A numeric value capturing the expected value of recall (share of
positive cases among those exhibiting the outcome). This value only needs to be
entered if `pi0` is left blank.}

\item{k}{A numeric value capturing the imbalance ratio (number of predicted positives
over predicted negatives) in the population from which the test set should be
drawn. Argument `k` can be left blank, but then `positive_share` must be specified.}

\item{positive_share}{A numeric value capturing the positive share (number of
predicted positives out of all observations) in the population from which the
test set should be drawn. Argument `positive_share` can be left blank, but
then `k` must specified.}

\item{external_k}{A numeric value capturing the imbalance ratio (number of predicted
positives over predicted negatives) in the population on which the recall estimate
is based. This only needs to be entered if `recall` is based on a population
that has a different imbalance than the test's population. The argument is
only required if `pi0` is not entered directly. If left blank, it will be
calculated internally using `external_positive_share`, if this is entered, or
assumed to be equal to the test population's imbalance.}

\item{external_positive_share}{A numeric value capturing the positive share
(number of predicted positives out of all observations) in the population on
which the recall estimate is based. This only needs to be entered if `recall`
is based on a population that has a different imbalance than the test's population.
The argument is only required if `pi0` is not entered directly. If left blank,
it will be calculated internally using `external_positive_share`, if this is
entered, or assumed to be equal to the test population's imbalance.}

\item{weight_se_f1}{A numeric value between 0 and 1 capturing how much to weigh
the SE for the F1 score in the objective function to be minimized by the calculator.
The default for `weight_se_f1` is `1`.}

\item{weight_se_rec}{A numeric value between 0 and 1 capturing how much to weigh
the SE for recall in the objective function to be minimized by the calculator.
The default for `weight_se_rec` is `0`.}

\item{weight_se_prec}{A numeric value between 0 and 1 capturing how much
to weigh the SE for precision in the objective function to be minimized by the
calculator. The default for `weight_se_prec` is `0`.}
}
\value{
A numeric value capturing how many positive observations should be sampled
given the parameter values and the objective function specified by the user.
}
\description{
This function estimates how many positives should optimally be sampled
}
\details{
This function can be used to estimate how many positive observations should be
included in the test set under efficient stratified random sampling (SRS). This
is the value that minimizes an objective function subject to a sample size constraint.
The objective function is a linear combination of the standard errors (SEs) of
precision, recall and F1 score, where weights can be specified by the user.
These standard errors (SEs) are based on analytic formulas for the
variance of the sampling distribution of the test set estimates of F1, precision
and recall under two-bin stratified sampling. Such variances are based on expected
parameter values, which need to be entered by the user. The value returns a
unique integer value that minimizes the expression subject to the constraint.
}
