% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_df.R
\docType{data}
\name{pop_df}
\alias{pop_df}
\title{Example data of a population from which to draw a test set}
\format{
A data frame with 20000 rows and 2 variables:
\describe{
  \item{id}{Identifier of each sampled observation}
  \item{score}{Predicted probaility of exhibiting the outcome}
}
}
\usage{
data(pop_df)
}
\description{
Fake data where each row corresponds to some observation in the population from
which the test set is to be sampled, and for which we know its predicted
probability of exhibiting the outcome of interest according to some classifier.
}
\examples{
data(pop_df)
}
\keyword{datasets}
