% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_df.R
\docType{data}
\name{sample_df}
\alias{sample_df}
\title{Example data of a test set}
\format{
A data frame with 1394 rows and 6 variables:
\describe{
  \item{id}{Identifier of each sampled observation}
  \item{score}{Predicted probaility of exhibiting the outcome}
  \item{strata}{Stratum of the sampled observation (based on score)}
  \item{Prob}{Sampling probability}
  \item{truth}{Annotated labels: whether the observation actually exhibits the outcome}
}
}
\usage{
data(sample_df)
}
\description{
Fake data with where each row corresponds to some observation in the annotated
test set, for which we know its stratum, sampling probability and true label.
}
\examples{
data(sample_df)
}
\keyword{datasets}
