% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_ci.R
\name{se_srs}
\alias{se_srs}
\title{Estimate standard errors for simple random sample}
\usage{
se_srs(tr0_pred0, tr1_pred0, tr0_pred1, tr1_pred1, N_sample)
}
\arguments{
\item{tr0_pred0}{A numeric value capturing the share of true negatives.}

\item{tr1_pred0}{A numeric value capturing the share of false negatives.}

\item{tr0_pred1}{A numeric value capturing the share of false positives.}

\item{tr1_pred1}{A numeric value capturing the share of true positives.}

\item{N_sample}{A numeric value capturing the total number size of the test set.}
}
\value{
A data.frame containing the analytic SEs for the F1 score, precision and recall.
These CIs are based on assuming the test set was drawn by simple random sampling:
the CIs should not be trusted if sampling was conducted in a different way.
}
\description{
This function analytically estimates SEs from confusion matrix drawn by simple random sampling
}
\details{
This function takes a confusion matrix and sample size as input, and produces
SEs with Wilson method for the F1 score, precision and recall. These SEs are
based on assuming the test set was drawn by simple random sampling: the CIs
should not be trusted if sampling was conducted in a different way.
}
\examples{
## enter the estimates for the confusion matrix cells and sample size
## get the SE analytically under assumption that test set is a SRS
se_srs(tr0_pred0 = 0.5, tr1_pred0 = 0.1, tr0_pred1 = 0.1, tr1_pred1 = 0.3, N_sample = 1000)
}
