% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_ci.R
\name{se_strat}
\alias{se_strat}
\title{Estimate standard errors for stratified random sample}
\usage{
se_strat(tr0_pred0, tr1_pred0, tr0_pred1, tr1_pred1, N_positive, N_negative)
}
\arguments{
\item{tr0_pred0}{A numeric value capturing the share of true negatives.}

\item{tr1_pred0}{A numeric value capturing the share of false negatives.}

\item{tr0_pred1}{A numeric value capturing the share of false positives.}

\item{tr1_pred1}{A numeric value capturing the share of true positives.}

\item{N_positive}{A numeric value capturing the number size of positive observations
sampled into the test set.}

\item{N_negative}{A numeric value capturing the number size of negative observations
sampled into the test set.}
}
\value{
A data.frame containing the analytic SEs for the F1 score, precision and
recall under two-bin stratification of positives and negatives. These SEs are not
valid if the test set was not drawn through stratified sampling (e.g SRS) or if
it was drawn through some other stratification regime (either not by two-bin
positives vs negatives stratification, or by stratifying on some other variable).
}
\description{
This function analytically estimates SEs from confusion matrix drawn by simple random sampling
}
\details{
This function takes a confusion matrix and sample size as input, and produces
SEs with Wilson method for the F1 score, precision and recall. These SEs are
based on assuming two-bin stratified random sampling, where the two bins are
positive and negative observations: the SEs should not be trusted if sampling
was conducted in a different way (either not by two-bin positives vs negatives
stratification, or by stratifying on some other variable).
}
\examples{
## enter the estimates for the test set confusion matrix cells and sample sizes
## get the SE analytically under assumption that test set is a two-bin
## stratified sample (positives and negatives)
se_strat(tr0_pred0 = 0.5, tr1_pred0 = 0.1, tr0_pred1 = 0.1, tr1_pred1 = 0.3,
         N_positive = 600, N_negative = 400)
}
