% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_fixer.R
\name{weight_fixer}
\alias{weight_fixer}
\title{Fix weights after incomplete annotations}
\usage{
weight_fixer(data, truth = "truth", strata = "strata", probs = NULL)
}
\arguments{
\item{data}{A data.frame containing the annotated test set.}

\item{truth}{A character value capturing the column name of the annotated labels.
Column should capture whether each observation exhibits the outcome (1/TRUE)
or not (0/FALSE). NAs should be entered for observations that were not annotated.}

\item{strata}{A character value capturing the column name of the sampling strata.
Column should contain factor or character values.}

\item{probs}{A character value capturing the column name of the sampling probabilities.
Default assumes equal sampling probability for all observations in the original sampling.}
}
\value{
A data.frame just like the input `data`, but with recomputed sampling
probabilities in column `Prob_new`.
}
\description{
This function estimates probability weights given a partially annotated test set
}
\details{
This function can be used on a test set drawn by stratified sampling that has
not been fully annotated, e.g. if human annotators stopped annotating halfway
through the annotation task. The function re-computes sampling probabilities
given the annotation was incomplete. This assumes that the comments that were
indeed annotated from each stratum are a SRS of the stratum.
}
\examples{
## assume that some observations were not annotated, such that their value in
## the column with annotated labels is NA. function will update the weights so
## estimates remain unbiased. Note this assumes that the annotated observations
## are a SRS of their respective stratum if the test set was drawn by stratified
## sampling, or a SRS of the entire population if the test set was drawn by SRS.
data(sample_df)
sample_df$truth[30:45] <- NA #code some labels as missing because not annotated
weight_fixer(data = sample_df, truth = "truth", strata = "strata", probs="Prob")
}
